+++ crates/gpui/src/platform/mac.rs
@@ -4,6 +4,7 @@ mod dispatcher;
 mod display;
 mod display_link;
 mod events;
+mod keyboard;
 mod screen_capture;
 
 #[cfg(not(feature = "macos-blade"))]
@@ -45,6 +46,7 @@ use std::{
 pub(crate) use dispatcher::*;
 pub(crate) use display::*;
 pub(crate) use display_link::*;
+pub(crate) use keyboard::*;
 pub(crate) use platform::*;
 pub(crate) use window::*;
